Multithreading (#7)
* Start working on multithreading for rendering part

* Implement correct (more or less) multithreading to start games together with application
Fully implement RenderTargetPanel control
Fix saving of gif images
Improve DrawingContext.cs
Extend EntityService.cs

* Make a lot improvements to image processing
Improve Measure and Arrange for Grid control
Rework loading of bitmap images
Make possibility for multiThreaded rendering for Game class inside UIApplication class
Made some improvements to AUML parser

* Rename image file types

* Improvements to IRawBitmap tto support saving to file

* Changes in Loading/Saving images
Improvements to Grid control measure/arrange logic
Improvements to rendering performance

* Improvements to DrawingContext class to use hash instead of direct comparing of input parameters

+++ Adamantium/Adamantium.Engine.Core/UidGenerator.cs
@@ -17,13 +17,13 @@ namespace Adamantium.Engine.Core
         /// Generate unique int64 number that can be used as UID
         /// </summary>
         /// <returns></returns>
-        public static Int64 Generate()
+        public static UInt128 Generate()
         {
             lock (syncObject)
             {
                 var randomData = new byte[length];
                 RandomGenerator.GetNonZeroBytes(randomData);
-                return BitConverter.ToInt64(randomData, 0);
+                return (UInt128)BitConverter.ToInt64(randomData, 0);
             }
         }
     }
